Update QuestStage
Quest stage
-Treasure [not Complete]
-Princess [not Complete]
-City transport [ not Complete]

+++ Assets/MyAsset/script/tileManager.cs
@@ -21,11 +21,12 @@ public class tileManager : MonoBehaviour {
 
     public static bool dungeonStage = false;
     public static bool bossStage = false;
+    public static bool QuestStage = false;
 
+    public static bool EndStage = false;
 
 	void Start () {                                                         //Spawn StarterPlatform when start Game
         playerPos = GameObject.Find("player").transform;
-
         allPlatformGame = new List<GameObject>();
 
         for (int i = 0; i < 6; i++)
@@ -43,11 +44,16 @@ public class tileManager : MonoBehaviour {
 	}
 	
 	void Update () {
-        
-        if (playerPos.position.z +arriveZone  >= spawnPlatformPoint && dungeonStage == false && bossStage == false) //allow front point player Check to The last edge platform
+        DistanceSpawnPlatform();
+	}
+
+    #region Platform from distance
+    void DistanceSpawnPlatform()
+    {
+        if (playerPos.position.z +arriveZone  >= spawnPlatformPoint && dungeonStage == false && bossStage == false && QuestStage == false && EndStage == false) //allow front point player Check to The last edge platform
         {
-                Spawnplatform(1);           // clone (normal platform)
-                DeleteOldplatform();        // delete (Old platform)
+            Spawnplatform(1);           // clone (normal platform)
+            DeleteOldplatform();        // delete (Old platform)
         }
 
         else if (playerPos.position.z +arriveZone  >= spawnPlatformPoint && dungeonStage == true) //allow front point player Check to The last edge platform
@@ -67,12 +73,22 @@ public class tileManager : MonoBehaviour {
                 bossGameplay.name = "Boss";
             }
         }
-	}
+        else if (playerPos.position.z +arriveZone  >= spawnPlatformPoint && dungeonStage == false && bossStage == false && QuestStage == true && EndStage == false) //allow front point player Check to The last edge platform
+        {
+            Debug.Log("Complete Stage");
+            Spawnplatform(0);
+            Spawnplatform(0);
+            Spawnplatform(3);           // clone (empty platform)
+            DeleteOldplatform();        // delete (Old platform)
+            EndStage = true;
+        }
+    }
+    #endregion
 
-    #region Spawn with TypePlatform
-    void Spawnplatform(int isRandom)
+    #region TypePlatform Function
+    void Spawnplatform(int typePlatform)
     {
-        if (isRandom == 0)                 
+        if (typePlatform == 0)                 
         {
             GameObject floor = Instantiate(platform[0], new Vector3(
                 starterplatform.position.x,
@@ -83,7 +99,7 @@ public class tileManager : MonoBehaviour {
             spawnPlatformPoint += sizePlatform;
             allPlatformGame.Add(floor);
         }
-        else if( isRandom == 1)
+        else if( typePlatform == 1)
         {
             int randomNormalplatform = Random.Range(1, 7);
             GameObject floor = Instantiate(platform[randomNormalplatform], new Vector3(
@@ -95,7 +111,7 @@ public class tileManager : MonoBehaviour {
                 spawnPlatformPoint += sizePlatform;
                 allPlatformGame.Add(floor);
         }
-        else if( isRandom == 2)
+        else if( typePlatform == 2)
         {
             /*int randomTrapPlatform = Random.Range(7, 8);*/                        //if want more TrapPlatform Check this
             GameObject trapFloor = Instantiate(platform[7], new Vector3(
@@ -106,6 +122,18 @@ public class tileManager : MonoBehaviour {
             trapFloor.transform.SetParent(this.transform);
             spawnPlatformPoint += sizePlatform;
             allPlatformGame.Add(trapFloor);
+        }
+        else if( typePlatform == 3)
+        {
+            GameObject specialFloor = Instantiate(platform[8], new Vector3(
+                starterplatform.position.x,
+                starterplatform.position.y,                                                                          //Random Trap platform
+                spawnPlatformPoint),
+                starterplatform.rotation) as GameObject;
+            specialFloor.name = "VictoryPlatform";
+            specialFloor.transform.SetParent(this.transform);
+            /*spawnPlatformPoint += sizePlatform;*/                 //was finish don't use
+            allPlatformGame.Add(specialFloor);
         }
     }
     void DeleteOldplatform()
@@ -118,12 +146,4 @@ public class tileManager : MonoBehaviour {
         }
     }
     #endregion
-
-    #region BossSpawn
-    void SpawnBoss()
-    {
-        
-    }
-    #endregion
-
 }
